requirements.txt
psycopg2==2.9.3
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
tzdata==2022.1
tzlocal==4.2